junit travis test
