+++ android/.travis.yml
@@ -1,0 +1,20 @@
+language: android
+android:
+  components:
+    - build-tools-23.0.1
+    - android-23
+
+    # Specify at least one system image,
+    # if you need to run emulator(s) during your tests
+    #- sys-img-armeabi-v7a-android-22
+    #- sys-img-x86-android-17
+
+# Emulator Management: Create, Start and Wait
+before_script:
+  - echo no | android create avd --force -n test -t android-22 --abi armeabi-v7a
+  - emulator -avd test -no-audio -no-window &
+  - android-wait-for-emulator
+  - adb shell input keyevent 82 &
+
+install: true
+script: - ./gradlew build connectedCheck
