+++ android/build.gradle
@@ -25,6 +25,14 @@ android {
     }
 }
 
+task copyResDirectoryToClasses(type: Copy){
+    from "${projectDir}/src/test/resources"
+    into "${buildDir}/intermediates/classes/test/debug/resources"
+    into "${buildDir}/intermediates/classes/test/release/resources"
+}
+
+tasks.assemble.dependsOn(tasks.copyResDirectoryToClasses)
+
 /* Fix for android test logging. Source: https://code.google.com/p/android/issues/detail?id=182307 */
 tasks.withType(com.android.build.gradle.internal.tasks.AndroidTestTask) { task ->
     task.doFirst {
@@ -36,11 +44,14 @@ tasks.withType(com.android.build.gradle.internal.tasks.AndroidTestTask) { task -
 }
 
 dependencies {
-    compile project(":core")
     compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile project(":core")
 
+    androidTestCompile 'com.android.support.test:runner:0.5'
+    testCompile 'org.hamcrest:hamcrest-all:1.3'
     testCompile 'junit:junit:4.12'
-    androidTestCompile 'com.android.support.test:runner:0.4'
-    androidTestCompile 'com.nanohttpd:nanohttpd:2.1.0'
-    androidTestCompile project(path: ':core', configuration: 'coreTestsConfiguration')
+    testCompile 'com.nanohttpd:nanohttpd:2.2.0'
+    testCompile 'org.nanohttpd:nanohttpd-nanolets:2.3.0'
+    testCompile 'org.mockito:mockito-core:2.0.46-beta'
+    testCompile project(path: ':core', configuration: 'coreTestsConfiguration')
 }
