+++ android/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Sep 02 13:18:37 EEST 2016
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.14.1-all.zip
