+++ android/src/test/java/io/ably/lib/test/realtime/RealtimeSuite.java
@@ -1,11 +1,8 @@
-package io.ably.lib.test.rest;
+package io.ably.lib.test.realtime;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.junit.runner.JUnitCore;
-import org.junit.runner.Result;
 import org.junit.runner.RunWith;
-import org.junit.runner.notification.Failure;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
@@ -13,20 +10,21 @@ import io.ably.lib.test.common.Setup;
 
 @RunWith(Suite.class)
 @SuiteClasses({
-	RestAppStatsTest.class,
-	RestInitTest.class,
-	RestTimeTest.class,
-	RestAuthTest.class,
-	RestTokenTest.class,
-	RestCapabilityTest.class,
-	RestChannelHistoryTest.class,
-	RestChannelPublishTest.class,
-	RestCryptoTest.class,
-	RestPresenceTest.class,
-	RestProxyTest.class,
-	HttpTest.class
+	EventEmitterTest.class,
+	RealtimeInitTest.class,
+	RealtimeConnectTest.class,
+	RealtimeConnectFailTest.class,
+	RealtimeChannelTest.class,
+	RealtimePresenceTest.class,
+	RealtimeMessageTest.class,
+	RealtimeResumeTest.class,
+	RealtimeRecoverTest.class,
+	RealtimeCryptoTest.class,
+	RealtimeCryptoMessageTest.class,
+	RealtimeChannelHistoryTest.class,
+	RealtimePresenceHistoryTest.class
 })
-public class RestSuite {
+public class RealtimeSuite {
 
 	@BeforeClass
 	public static void setUpBeforeClass() throws Exception {
