+++ android/src/test/resources/local/testAppSpec.json
@@ -1,0 +1,74 @@
+{
+	"keys": [
+		{},
+		{
+			"capability": "{\"testchannel\":[\"publish\"]}"
+		},
+		{
+			"capability": "{\"*\":[\"subscribe\"],\"canpublish:*\":[\"publish\"],\"canpublish:andpresence\":[\"presence\",\"publish\"]}"
+		},
+		{
+			"capability": "{\"candoall\":[\"*\"]}"
+		},
+		{
+			"capability": "{\"channel0\":[\"publish\"],\"channel1\":[\"publish\"],\"channel2\":[\"publish\",\"subscribe\"],\"channel3\":[\"subscribe\"],\"channel4\":[\"presence\",\"publish\",\"subscribe\"],\"channel5\":[\"presence\"],\"channel6\":[\"*\"]}"
+		},
+		{
+			"privileged": true,
+			"capability": "{\"channel0\":[\"publish\"],\"channel1\":[\"publish\"],\"channel2\":[\"publish\",\"subscribe\"],\"channel3\":[\"subscribe\"],\"channel4\":[\"presence\",\"publish\",\"subscribe\"],\"channel5\":[\"presence\"],\"channel6\":[\"*\"]}"
+		}
+	],
+	"namespaces": [
+		{
+			"id": "persisted",
+			"persisted": true
+		}
+	],
+	"channels": [
+		{
+			"name": "restpresence_notpersisted",
+			"presence": [
+				{
+					"clientId": "client_string",
+					"data": "This is a string data payload"
+				}
+			]
+		},
+		{
+			"name": "persisted:restpresence_persisted",
+			"presence": [
+				{
+					"clientId": "client_string_0",
+					"data": "This is a string data payload"
+				}
+			]
+		},
+		{
+			"name": "persisted:restpresence_persisted",
+			"presence": [
+				{
+					"clientId": "client_string_1",
+					"data": "This is a string data payload"
+				}
+			]
+		},
+		{
+			"name": "persisted:restpresence_persisted",
+			"presence": [
+				{
+					"clientId": "client_string_2",
+					"data": "This is a string data payload"
+				}
+			]
+		},
+		{
+			"name": "persisted:restpresence_persisted",
+			"presence": [
+				{
+					"clientId": "client_string_3",
+					"data": "This is a string data payload"
+				}
+			]
+		}
+	]
+}
