+++ android/src/test/resources/local/testAppSpec.json.src
@@ -1,0 +1,83 @@
+{
+	keys: [
+		{},  /* key 0, blanket */
+		{    /* key 1, specific channel and ops */
+			capability: JSON.stringify({
+				testchannel:['publish']
+			})
+		},
+		{   /* key 2, wildcard channel spec */
+			capability: JSON.stringify({
+				'*':['subscribe'],
+				'canpublish:*':['publish'],
+				'canpublish:andpresence':['presence', 'publish']
+			})
+		},
+		{   /* key 3,  wildcard ops spec */
+			capability: JSON.stringify({
+				'candoall':['*']
+			})
+		},
+		{   /* key 4, multiple resources */
+			capability: JSON.stringify({
+				channel0:['publish'],
+				channel1:['publish'],
+				channel2:['publish', 'subscribe'],
+				channel3:['subscribe'],
+				channel4:['presence', 'publish', 'subscribe'],
+				channel5:['presence'],
+				channel6:['*']
+			})
+		},
+		{   /* key 5, has wildcard clientId */
+			privileged: true,
+			capability: JSON.stringify({
+				channel0:['publish'],
+				channel1:['publish'],
+				channel2:['publish', 'subscribe'],
+				channel3:['subscribe'],
+				channel4:['presence', 'publish', 'subscribe'],
+				channel5:['presence'],
+				channel6:['*']
+			})
+		}
+	],
+	namespaces: [
+		{	/* persisted namespace, persisted */
+			id: 'persisted',
+			persisted: true
+		}
+	],
+	channels: [
+		{	/* non-persisted */
+			name: 'restpresence_notpersisted',
+			presence: [
+				{	clientId: 'client_string', data: 'This is a string data payload' }
+			]
+		},
+		{	/* persisted */
+			name: 'persisted:restpresence_persisted',
+			presence: [
+				{	clientId: 'client_string_0', data: 'This is a string data payload' }
+			]
+		},
+		{	/* persisted */
+			name: 'persisted:restpresence_persisted',
+			presence: [
+				{	clientId: 'client_string_1', data: 'This is a string data payload' }
+			]
+		},
+		{	/* persisted */
+			name: 'persisted:restpresence_persisted',
+			presence: [
+				{	clientId: 'client_string_2', data: 'This is a string data payload' }
+			]
+		},
+		{	/* persisted */
+			name: 'persisted:restpresence_persisted',
+			presence: [
+				{	clientId: 'client_string_3', data: 'This is a string data payload' }
+			]
+		},
+	]
+}
