+++ core/src/test/java/io/ably/lib/test/common/Setup.java
@@ -1,14 +1,14 @@
 package io.ably.lib.test.common;
 
+import com.google.gson.Gson;
+
 import java.io.IOException;
 import java.util.Collection;
 import java.util.Date;
 
-import com.google.gson.Gson;
-
-import io.ably.lib.http.HttpUtils;
 import io.ably.lib.http.Http.JSONRequestBody;
 import io.ably.lib.http.Http.ResponseHandler;
+import io.ably.lib.http.HttpUtils;
 import io.ably.lib.rest.AblyRest;
 import io.ably.lib.types.AblyException;
 import io.ably.lib.types.ClientOptions;
@@ -200,7 +200,7 @@ public class Setup {
 				claz = (Class<? extends ResourceLoader>) Class.forName("io.ably.lib.test.common.jre.JreResourceLoader");
 			} catch(ClassNotFoundException cnfe) {
 				try {
-					claz = (Class<? extends ResourceLoader>) Class.forName("io.ably.lib.test.android.AssetResourceLoader");
+					claz = (Class<? extends ResourceLoader>) Class.forName("io.ably.lib.test.common.android.AssetResourceLoader");
 				} catch(ClassNotFoundException cnfe2) {
 					System.err.println("Unable to instance any known ResourceLoader class");
 				}
